Revert "feat: auth0 backend authorization flow with permissions retrieval from MGMT API as fallback strat., plus provisional cache feats for access token and permission storage"
