--- backend/.dockerignore
@@ -1,8 +1,0 @@
-target/
-.idea/
-*.iml
-*.log
-*.md
-.git/
-.gitignore
-.DS_Store
