--- backend/docker-compose.yml
@@ -1,23 +1,0 @@
-services:
-  scouts-project:
-    env_file:
-      - .env
-    image: scouts_build:dev
-    container_name: scouts_dev
-    build:
-      context: ./
-      dockerfile: Dockerfile
-      target: dev
-    init: true
-    volumes:
-      - type: bind
-        source: ./
-        target: /app
-    ports:
-      - "8080:8080"
-    restart: unless-stopped
-    networks:
-      - backend
-
-networks:
-  backend:
