+++ backend/src/main/java/uao/edu/co/scouts_project/ScoutsProjectApplication.java
@@ -4,10 +4,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 import io.github.cdimascio.dotenv.Dotenv;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
-@EnableScheduling
 public class ScoutsProjectApplication {
 
 	public static void main(String[] args) {
