+++ backend/src/main/resources/application.properties
@@ -15,22 +15,3 @@ spring.jpa.show-sql=${SHOW_SQL:false}
 
 # ==== Servidor ====
 server.port=${SERVER_PORT:8080}
-
-
-# -- Auth0 authz flow
-debug=true
-logging.level.uao.edu.co.scouts_project=DEBUG
-logging.level.org.springframework.security=DEBUG
-
-okta.oauth2.issuer=${OKTA_OAUTH2_ISSUER}
-okta.oauth2.audience=${OKTA_OAUTH2_AUDIENCE}
-
-auth0.domain=${AUTH0_DOMAIN}
-auth0.clientId=${AUTH0_CLIENT_ID}
-auth0.clientSecret=${AUTH0_CLIENT_SECRET}
-auth0.audience=${AUTH0_AUDIENCE}
-
-auth0.management.eager-init=${AUTH0_MANAGEMENT_EAGER_INIT:true}
-auth0.management.background-refresh.enabled=${AUTH0_MANAGEMENT_BACKGROUND_REFRESH_ENABLED:true}
-auth0.management.refresh-check-interval=${AUTH0_MANAGEMENT_REFRESH_CHECK_INTERVAL:300000}
-authz.cache.ttl-millis=${AUTHZ_CACHE_TTL_MILLIS:300000}
