+++ backend/src/test/resources/application-test.properties
@@ -21,16 +21,3 @@ logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
 
 # Test specific settings
 spring.test.database.replace=none
-
-okta.oauth2.issuer=https://dummy.local.dev/oauth2/default
-okta.oauth2.audience=api://default-test
-
-auth0.domain=dev-xxxxx12345.us.auth0.com
-auth0.clientId=FAKE1234567890CLIENTID
-auth0.clientSecret=FAKE1234567890CLIENTSECRET
-auth0.audience=https://dev-xxxxx12345.us.auth0.com/api/v2/
-
-auth0.management.eager-init=true
-auth0.management.background-refresh.enabled=true
-auth0.management.refresh-check-interval=300000
-authz.cache.ttl-millis=300000
